fixed some dependencies in wrong headers
